third_party/blink/web_tests/external/wpt/css/css-align/blocks/align-content-block-007-expected.txt

This is a testharness.js-based test.
[FAIL] .test 5: last baseline
  assert_equals: \n<div class="test" style="align-content: last baseline" title="last baseline">\n    <div class="float" data-offset-y="35">FLT</div>\n    <div class="in-flow" data-offset-y="45"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="45">FLT</div>\n      <span class="label">LAST BASELINE</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 35 but got 40
[FAIL] .test 9: unsafe center
  assert_equals: \n<div class="test" style="align-content: unsafe center" title="unsafe center">\n    <div class="float" data-offset-y="25">FLT</div>\n    <div class="in-flow" data-offset-y="35"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="35">FLT</div>\n      <span class="label">UNSAFE CENTER</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 25 but got 35
[FAIL] .test 10: unsafe end
  assert_equals: \n<div class="test" style="align-content: unsafe end" title="unsafe end">\n    <div class="float" data-offset-y="15">FLT</div>\n    <div class="in-flow" data-offset-y="25"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="25">FLT</div>\n      <span class="label">UNSAFE END</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 15 but got 35
Harness: the test ran to completion.
